mirrorprobes: removing mirror pod probes (PROJQUAY-2226)
Following feedback there is now ay of getting the health of the mirror
component. This PR removes the checks, essentially rolling back a change
made by https://github.com/quay/quay-operator/pull/427.

As soon as we have a proper way of checking for the mirror health we
will need to re-enable these.