Removing replicas property from kustomize
If replicas are set we ended up competing with the HPA during the
reconcile as we set them back to what has been defined.

By removing the replicas property we just leave it alone to whatever HPA
decided to set it. If no HPA is configured then replicas will be set
automatically to 1 (this is the default behavior).