refactor(build-workflow): infer image tag from branch name (PROJQUAY-2612) (#556)
This change ties the build workflow to the name of the branches it runs against.
In a future where downstream releases are fully (or nearly) automated, we should probably tag the redhat-* branches for everything we want built, alpha, release candidates, etc. When this is done, this workflow should be updated to get triggered when new tags are created.