chore: automate downstream manifest generation
make prepare-downstream-release RELEASE=3.6.2 REPLACES=3.6.1 QUAY_RELEASE=3.6.2 CLAIR_RELEASE=3.6.2 CHANNEL=stable-3.6

will create a cluster service version virtually identical to the one
located in bundle/downstream.

note that we're also replacing `QUAY_DEFAULT_BRANDING` value from
'upstream' to 'redhat'. it seems the operator knows how to tell Quay
which is the correct logo to use. this is also done in the CSV itself,
and is still done because I'm actually not sure if the mechanism behind
QUAY_DEFAULT_BRANDING can be trusted.