controller: update condition reason when upgrade job crashes
the current code will only update the condition reason when the job
fails, but if the job crashes the reconcile loop will stay stuck in the
migration in progress state and never recreate the upgrade job.
this fixes the problem by always updating the condition reason to
migrations failed after having checked that the job is neither active
nor succeeded.

relates to PROJQUAY-1812