clair: limit db connections pool (PROJQUAY-2659)
Restrict the max number of open connections to the clair postgres. If
HPA is managed we have control over the max number of pods so we can
calculate the max number of connections per pod. If HPA is not managed
then we can't properly estimate the max number of pods so we decrease
the connection pool size.