[redhat-3.13] pgmigration: fixing managed clair postgres migration error (PROJQUAY-8048) (#976)
Clair and Quay postgres migrations were controlled by a single env variable. It was checking that the postgres version of new clair-postgres pods i.e postgres-15 was not matching the expected postgres version i.e. postgres-13. This retriggered the upgrade process making the postgres upgrade job stuck in the loop.

This change separates the clair postgres migration component and rectify the logic of validating the upgrade job completion.

---------

Co-authored-by: Shubhra Deshpande <shubhrajayant+github.com>